Multiple fixes for building with Qt6 on Windows
This was tested with the experimental vcpkg Qt6 port:
https://github.com/microsoft/vcpkg/pull/14333
as well as with the official Qt6 MSVC binary distribution

Working with upstream on the vcpkg port of Qt6 is the priority since
that is what will be used in CI.

What works:

- Dynamic and static Debug and Release builds on MSVC with vcpkg.
In all 4 cases, installation works correctly, as well as running from the
build directory right after the build.

- Dynamic Release and Debug builds on MSVC without vcpkg (using zlib,
  boost and libtorrent compiled from source + qt and openssl precompiled
  binaries from the official MSVC Qt distribution). Again, running from
  the build directory and installing works.

Current issues when building with the vcpkg Qt6 port:

- The problem with CMAKE_FIND_PACKAGE_PREFER_CONFIG (marked with a TODO)
has been reported to Qt: https://bugreports.qt.io/browse/QTBUG-93604
The workaround is trivial (just commented out a single line) and
harmless for now.

What was not tested/touched:

- Building with MinGW
- Static builds without vcpkg

Minor changes:

- Bumped minimum Qt version to 6.1.0
- Bumped minimum CMake version to 3.20 due to usage of cmake_path()
- The Qt6 Tools component is required for the Qt6::windeployqt target